feat(core): add token validator registry
