+++ .changeset/token-validator-registry.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': minor
+---
+
+add token validator registry for plugin extensibility
