+++ src/core/token-validators/typography.ts
@@ -1,0 +1,38 @@
+import type { TokenValidator } from './index.js';
+import { isRecord } from './utils.js';
+import validateFontFamily from './fontFamily.js';
+import validateDimension from './dimension.js';
+import validateFontWeight from './fontWeight.js';
+import validateNumber from './number.js';
+
+const validateTypography: TokenValidator = (value, path, tokenMap) => {
+  if (!isRecord(value)) {
+    throw new Error(`Token ${path} has invalid typography value`);
+  }
+  const allowed = new Set([
+    'fontFamily',
+    'fontSize',
+    'fontWeight',
+    'lineHeight',
+  ]);
+  for (const key of Object.keys(value)) {
+    if (!allowed.has(key)) {
+      throw new Error(`Token ${path} has invalid typography value`);
+    }
+  }
+  const { fontFamily, fontSize, fontWeight, lineHeight } = value;
+  if (
+    fontFamily === undefined ||
+    fontSize === undefined ||
+    fontWeight === undefined ||
+    lineHeight === undefined
+  ) {
+    throw new Error(`Token ${path} has invalid typography value`);
+  }
+  validateFontFamily(fontFamily, `${path}.fontFamily`, tokenMap);
+  validateDimension(fontSize, `${path}.fontSize`, tokenMap);
+  validateFontWeight(fontWeight, `${path}.fontWeight`, tokenMap);
+  validateNumber(lineHeight, `${path}.lineHeight`, tokenMap);
+};
+
+export default validateTypography;
